Fix error when resolving interface in federation
I encountered an error on my application while trying to resolve an interface from a reference resolver in a federated graph. I was able to replicate the issue in the apollo-federation example.

The first thing I did was update the example graphs to federation v2. Then I turned Product into an interface and created two subtypes for it.

I then used the example graphql and got my error.

If I changed the `resolveProductReference` to be synchronous the issue went away, but in my case I need it to be async.

Thanks to the stacktrace I was able to track it down to the instance in the resolveType function being a promise in this scenario. If we await it, everything looks good.