Add helpful comment about unique id in subscriptions resolver
The id is needed as a unique identifier for the Apollo Client cache. However, if the mongodb database har multiple instances, the naive approach in the documentation will be an issue since the id is only stored in-memory in each mongodb instance. Thus, there will be duplicates. 

Adding a comment about what the id is for gives the developer a chance to make an informed decision.